+++ packages/wordpress-kit/src/lib/tailwindcssPlugin/Config.ts
@@ -48,6 +48,8 @@ export const mergeToConfig: Config = {
 		'.wp-block-audio',
 		'.wp-block-cover',
 		'.wp-block-media-text',
+		'.wp-block-buttons',
+		'.wp-block-file',
 	],
 	...proseOverride,
 };