+++ packages/wordpress-kit/src/lib/tailwindcssPlugin/components/index.ts
@@ -6,6 +6,8 @@ import { GalleryComponent } from './Gallery';
 import { AudioComponent } from './Audio';
 import { CoverComponent } from './Cover';
 import { mediaAndText } from './MediaAndText';
+import { ButtonsComponent } from './Buttons';
+import { FileMediaComponent } from './FileMedia';
 
 export {
 	ImageComponent,
@@ -16,4 +18,6 @@ export {
 	AudioComponent,
 	CoverComponent,
 	mediaAndText,
+	ButtonsComponent,
+	FileMediaComponent,
 };