+++ packages/wordpress-kit/src/lib/tailwindcssPlugin/index.ts
@@ -8,6 +8,8 @@ import {
 	AudioComponent,
 	CoverComponent,
 	mediaAndText,
+	ButtonsComponent,
+	FileMediaComponent,
 } from './components';
 import { mergeToConfig } from './Config';
 import { ColorUtilities, FontsUtilities } from './utilities';
@@ -21,9 +23,7 @@ export default plugin(function ({ addUtilities, theme, addComponents }) {
 
 	const quoteUtilities = Quote;
 
-	const pullQuote = PullQuoteComponent({
-		quoteSize: theme('fontSize.4xl', '2.5rem'),
-	});
+	const pullQuote = PullQuoteComponent;
 
 	const table = TableComponent({
 		alignFull: {
@@ -48,6 +48,14 @@ export default plugin(function ({ addUtilities, theme, addComponents }) {
 		alignFull: { minWidth: theme('screen.xl', '1280px') },
 	});
 
+	const buttons = ButtonsComponent({
+		defaultColor: color.getColorByName('primary'),
+	});
+
+	const fileMedia = FileMediaComponent({
+		defaultColor: color.getColorByName('primary'),
+	});
+
 	addUtilities([
 		color.getBackgroundUtilities(),
 		color.getBorderColorUtilities(),
@@ -59,5 +67,15 @@ export default plugin(function ({ addUtilities, theme, addComponents }) {
 		quoteUtilities,
 	]);
 
-	addComponents([table, image, pullQuote, gallery, audio, cover, mediaAndText]);
+	addComponents([
+		table,
+		image,
+		pullQuote,
+		gallery,
+		audio,
+		cover,
+		mediaAndText,
+		buttons,
+		fileMedia,
+	]);
 }, mergeToConfig);