+++ packages/wordpress-kit/src/lib/tailwindcssPlugin/utilities/Colors.ts
@@ -100,12 +100,15 @@ export class ColorUtilities {
 			(acc, color) => ({
 				...acc,
 				[`.has-${color.name}-color`]: {
-					color: this.getColor(color),
+					color: `${this.getColor(color)} !important`,
 				},
 			}),
 			{
 				'.has-text-color': {
 					strong: {
+						color: 'inherit !important',
+					},
+					code: {
 						color: 'inherit !important',
 					},
 				},
@@ -223,4 +226,8 @@ export class ColorUtilities {
 			}),
 			{},
 		);
+	getColorByName = (colorName: string) => {
+		const color = this.findColor(colorName);
+		return this.getColor(color);
+	};
 }