+++ starters/next-wordpress-starter/components/grid.jsx
@@ -3,29 +3,9 @@ import Image from 'next/image';
 import { IMAGE_URL } from '../lib/constants';
 import { getUrlPath } from '../lib/getUrlPath';
 
-export const withGrid = (Component) => {
-	const GridedComponent = ({ data, ...props }) => {
-		if (!data || !data.length) {
-			return props.contentType ? (
-				<h2 className="text-xl text-center mt-14">
-					No {props.contentType} found 🏜
-				</h2>
-			) : null;
-		}
-
-		return (
-			<div
-				className={`mt-12 grid gap-5 max-w-content mx-auto lg:grid-cols-3 lg:max-w-screen-lg`}
-			>
-				{data.map((content, i) => {
-					return <Component key={i} content={content} {...props} />;
-				})}
-			</div>
-		);
-	};
-
-	return GridedComponent;
-};
+const GradientPlaceholder = () => (
+	<div className="w-full h-full bg-gradient-to-b from-blue-100 to-blue-500" />
+);
 
 export const PostGridItem = ({ content: post }) => {
 	const imgSrc = getUrlPath(post?.featuredImage?.node?.sourceUrl);
@@ -44,12 +24,7 @@ export const PostGridItem = ({ content: post }) => {
 								alt={altText}
 							/>
 						) : (
-							<Image
-								src="/pantheon.png"
-								alt="Pantheon Logo"
-								layout="fill"
-								className="bg-black"
-							/>
+							<GradientPlaceholder />
 						)}
 					</div>
 					<h2 className="my-4 mx-6 text-xl leading-7 font-semibold text-gray-900">
@@ -66,7 +41,7 @@ export const PageGridItem = ({ content: page }) => {
 	const altText = page?.featuredImage?.node.altText || page.title;
 
 	return (
-		<Link passHref href={`/pages${page.uri}`}>
+		<Link passHref href={`/pages/${page.uri}`}>
 			<a>
 				<div className="flex flex-col rounded-lg shadow-lg overflow-hidden cursor-pointer border-2 h-full hover:border-indigo-500">
 					<div className="flex-shrink-0 relative h-40">
@@ -78,12 +53,7 @@ export const PageGridItem = ({ content: page }) => {
 								alt={altText}
 							/>
 						) : (
-							<Image
-								src="/pantheon.png"
-								alt="Pantheon Logo"
-								layout="fill"
-								className="bg-black"
-							/>
+							<GradientPlaceholder />
 						)}
 					</div>
 					<h2 className="my-4 mx-6 text-xl leading-7 font-semibold text-gray-900">