+++ starters/next-wordpress-starter/pages/_app.jsx
@@ -1,4 +1,5 @@
 import '../styles/globals.css';
+import '@pantheon-systems/nextjs-kit/style.css';
 
 function MyApp({ Component, pageProps }) {
 	return <Component {...pageProps} />;