+++ packages/drupal-kit/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @pantheon-systems/drupal-kit
 
+## 3.1.0-canary.0
+
+### Minor Changes
+
+- 74ce69e: Remove unused dependencies and clean up versions in package.json
+
+### Patch Changes
+
+- aea42f5: Fix types field in package.json
+
+## 3.0.1
+
+### Patch Changes
+
+- 9dbbc6b: Updated dist bundle to include generated d.ts files
+
+## 3.0.1-canary.0
+
+### Patch Changes
+
+- 9dbbc6b: Updated dist bundle to include generated d.ts files
+
 ## 3.0.0
 
 ### Major Changes