+++ starters/gatsby-wordpress-starter/CHANGELOG.md
@@ -1,5 +1,39 @@
 # @pantheon-systems/gatsby-wordpress-starter
 
+## 2.3.0-canary.1
+
+### Minor Changes
+
+- f99244a: Refactored to remove react-helmet and use the gatsby Head
+
+## 2.3.0-canary.0
+
+### Minor Changes
+
+- 74ce69e: Upgraded `gatsby` to v4.24.0 and bumped gatsby related plugins
+
+## 2.2.0
+
+### Minor Changes
+
+- 2bfe07a: Remove pantheon logo for cards that don't have featured image, and
+  use a generic gradient instead
+
+### Patch Changes
+
+- c634786: Add .nvmrc file
+
+## 2.2.0-canary.0
+
+### Minor Changes
+
+- 2bfe07a: Remove pantheon logo for cards that don't have featured image, and
+  use a generic gradient instead
+
+### Patch Changes
+
+- c634786: Add .nvmrc file
+
 ## 2.1.0
 
 ### Minor Changes